sign_up_window updated
A series of changes were made
